[LibOS] Split `shim_mount` and `shim_fs`
This change replaces most usages of `shim_mount` with `shim_fs`.
Graphene uses the `shim_mount` structure to represent a mounted
filesystem, but in most cases we're only interested in the function
table.

The main reason for this change is special treatment of sockets and
FIFOs: before, we replaced the `fs` field (actually `shim_mount`) for
their dentries, so that they are served by a different filesystem.
However, that loses information about which part of the tree the dentry
actually is in.

The new solution is to store both `shim_mount` and `shim_fs` for a
dentry. This follows Linux, which stores both `super_block` and
`inode_iperations` for an inode, and replaces the latter in case of
sockets and FIFOs. This will allow me to refactor the way mounted
filesystems are represented.

As a bonus, this change removes some hacks like unwarranted usage of
`*_builtin_fs` and storing filesystem type separately for checkpointing
purposes.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>