[LibOS] Rewrite string-based files
This change decouples the two usages of "str" filesystem: "pseudo"
filesystems, and `tmpfs`. The difference between them is that in
"pseudo" filesystems, the file data is stored directly in handle.

Instead of `str_*` functions, there is a simpler set of `mem_file_*`
functions that do not operate on handles or dentries, just on its own
buffer.

The decoupling makes both use-cases simpler: there is no structure
"inheritance" and no extra layer of reference counting.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>