[LibOS] Define and use `file_off_t` for file size/offset
The `off_t` type can be 32-bit, so we define a `file_off_t` type that is
guaranteed to be 64-bit.

(Linux kernel uses the 64-bit `loff_t` type, but we define our own for
clarity and to avoid depending on libc types).

This commit also introduces the `generic_seek` function, to ensure that
all `seek` operations check for errors in the same way. It also fixes
the return value of `lseek` on directories.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>